Began splitting script into modules
