+++ index.html
@@ -68,6 +68,6 @@
             </div>
             </div>
             <button type='button' id='resetButton'>Reset</button>
-        <script src='script.js'></script>
+        <script type='module' src='script.js'></script>
         </body>
     </html>
